Feature/#89 creation test pipeline

+++ app/database/database.py
@@ -1,19 +1,29 @@
 from sqlalchemy import create_engine, Column, Integer, Text
 from sqlalchemy.orm import sessionmaker, declarative_base
+from sqlalchemy.pool import StaticPool
 import os
 
 DATABASE_URL = os.getenv('DATABASE_URL')
 
 if not DATABASE_URL:
-  DATABASE_URL = 'sqlite:///./test.db'
-  print(f'AVISO: A variável de ambiente DATABASE_URL não foi definida. Usando "{DATABASE_URL}" como padrão.')
+  # Using a temporary path to a database during tests
+  DATABASE_URL = f'sqlite:///{os.getcwd()}/test.db'
+  print(f'WARN: The DATABASE_URL environment variable was not found. Using "{DATABASE_URL}" as the default url.')
 
 connect_args = {'check_same_thread': False} if 'sqlite' in DATABASE_URL else {}
 
-engine = create_engine(
-  DATABASE_URL,
-  connect_args=connect_args,
-)
+# If the databse is stored in memory (sqlite) don't try to remove the file
+if 'sqlite' in DATABASE_URL and ':memory:' in DATABASE_URL:
+  engine = create_engine(
+    DATABASE_URL,
+    connect_args={'check_same_thread': False},
+    poolclass=StaticPool,
+  )
+else:
+  engine = create_engine(
+    DATABASE_URL,
+    connect_args=connect_args,
+  )
 
 SessionLocal = sessionmaker(
   autocommit=False,
@@ -42,23 +52,15 @@ def get_db():
 
 
 def create_tables():
-  if 'sqlite' in DATABASE_URL and os.path.exists('./test.db'):
-    print('Excluindo arquivo de banco de dados existente: ./test.db')
+  # Will only remove the .db file if the database is sotred on memory (sqlite)
+  if 'sqlite' in DATABASE_URL and os.path.exists('./test.db') and ':memory:' not in DATABASE_URL:
+    print('Removing .db file: ./test.db')
     os.remove('./test.db')
-
-  # try:
-  #   with engine.connect() as connection:
-  #     connection.execute(text('CREATE SCHEMA IF NOT EXISTS session_manager'))
-  #     connection.commit()
-  #     print('Schema "session_manager" criado ou já existe.')
-  # except Exception as e:
-  #   print(f'Erro ao criar schema "session_manager": {e}')
-
   try:
     Base.metadata.create_all(bind=engine)
-    print('Tabelas do banco de dados criadas (ou já existem).')
+    print('Database tables created.')
   except Exception as e:
-    print(f'Erro ao criar tabelas: {e}')
+    print(f'ERROR: Error while creating database tables: {e}')
     raise
